chore: organize file structure
