+++ src/components/JobCard/JobBadge.tsx
