+++ src/components/JobCard/JobCard.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@chakra-ui/react";
 import { IJob } from "@/database/jobSchema";
-import JobBadge from "@/components/JobBadge";
-import JobCardInformation from "@/components/JobCardInformation";
+import JobBadge from "@/components/JobCard/JobBadge";
+import JobCardInformation from "@/components/JobCard/JobCardInformation";
 
 interface JobCardProps {
   job: IJob;
